Fix golden tests and drop pretty-simple
For some reason (probably changes in unordered-containers), golden tests
were broken on Stack because HashMap outputed its Show representation
with elements shuffled around in a different permutation.

Now the golden tests do not merely compare the textual output, but the
golden text is `read` and compared with `(==)` to the parsed value instead.