Implement GDB session management
This implementation heavily relies on Haskell threads are a special
thread (the "manager") is always listening for messages coming from GDB.
These are passed to other threads with 'MVar's. The manager may block if
handlers do not immediately take the 'MVar's, but the rationale is that
non-blocking behavior can always be implemented by the user by spawning
other threads.

TODO: This implementation is not exception-safe at all. It is not clear
what happens to handler threads that keep waiting indefinitely on empty
'MVar's when they are replaced with other handlers or when the GDB
session is closed. In Italian we would say "vanno a farsi fottere".

TODO: Implement tests